requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
boto3==1.34.69
gunicorn==21.2.0
python-dotenv==1.0.1
